+++ foxes/engines/__init__.py
@@ -4,6 +4,7 @@ from .numpy import NumpyEngine
 from .single import SingleChunkEngine
 from .futures import ThreadsEngine, ProcessEngine
 from .mpi import MPIEngine
+from .ray import RayEngine
 
 from .dask import (
     DaskBaseEngine,